requirements.txt
azure-functions
azure-storage-blob
pymongo>=4.6.0
dnspython>=2.4.0
openai>=1.0.0
pypdf
python-dotenv
numpy
groq
